% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictlm-fit.R
\name{predictlm_fit}
\alias{predictlm_fit}
\title{Fit predictlm Tabular Foundation Model}
\usage{
predictlm_fit(X, y, checkpoint = "mini", problem_type = "classification", ...)
}
\arguments{
\item{X}{x training data}

\item{y}{y training data}

\item{checkpoint}{"mini" or "base"}

\item{problem_type}{regression or classification}

\item{...}{for predictlm additional parameters}
}
\value{
a predictlm fit
}
\description{
Fit predictlm Tabular Foundation Model
}
