% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictlm-pred.R
\name{predictlm_predict}
\alias{predictlm_predict}
\title{PredictLM Predictions}
\usage{
predictlm_predict(fit, data, type = "raw", ...)
}
\arguments{
\item{fit}{a fitted model}

\item{data}{a dataset}

\item{type}{'raw'}

\item{...}{for passing parameters to the predict method}
}
\value{
predictions
}
\description{
PredictLM Predictions
}
